fix: authOptionsを別ファイルに切り出す
