+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,63 +1,5 @@
-import * as jsonwebtoken from 'jsonwebtoken';
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import { JWT } from 'next-auth/jwt';
-import GithubProvider from 'next-auth/providers/github';
-import GoogleProvider from 'next-auth/providers/google';
-import { HasuraAdapter } from 'next-auth-hasura-adapter';
-
-const authOptions: NextAuthOptions = {
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID as string,
-      clientSecret: process.env.GITHUB_SECRET as string,
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_ID as string,
-      clientSecret: process.env.GOOGLE_SECRET as string,
-    }),
-  ],
-  adapter: HasuraAdapter({
-    endpoint: process.env.NEXT_PUBLIC_GRAPHQL_ENDPOINT!,
-    adminSecret: process.env.HASURA_ADMIN_SECRET!,
-  }),
-  // pages: {
-  //   signIn: '/',
-  // },
-  session: { strategy: 'jwt' },
-  jwt: {
-    encode: ({ secret, token }) => {
-      const encodedToken = jsonwebtoken.sign(token!, secret, {
-        algorithm: 'HS256',
-      });
-      return encodedToken;
-    },
-    decode: async ({ secret, token }) => {
-      const decodedToken = jsonwebtoken.verify(token!, secret, {
-        algorithms: ['HS256'],
-      });
-      return decodedToken as JWT;
-    },
-  },
-  callbacks: {
-    async jwt({ token }) {
-      return {
-        ...token,
-        'https://hasura.io/jwt/claims': {
-          'x-hasura-allowed-roles': ['user'],
-          'x-hasura-default-role': 'user',
-          'x-hasura-role': 'user',
-          'x-hasura-user-id': token.sub,
-        },
-      };
-    },
-    session: async ({ session, token }) => {
-      if (session?.user) {
-        session.user.id = token.sub!;
-      }
-      return session;
-    },
-  },
-};
+import { authOptions } from '@utils/authOptions';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
